# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Oss对象存储",
        "title": "Oss对象存储",
        "termsOfService": "https://elvesyuki.com",
        "contact": {
            "name": "luohuan",
            "url": "https://elvesyuki.com",
            "email": "1026770043@qq.com"
        },
        "version": "1.0"
    },
    "host": "localhost:9091",
    "paths": {
        "/v1/oss/multiple/upload": {
            "post": {
                "description": "Oss多文件上传接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oss文件上传接口"
                ],
                "summary": "Oss多文件上传接口",
                "parameters": [
                    {
                        "type": "file",
                        "description": "对象文件数组",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传编码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresult.Response"
                        }
                    }
                }
            }
        },
        "/v1/oss/single/upload": {
            "post": {
                "description": "Oss单个文件上传接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oss文件上传接口"
                ],
                "summary": "Oss单个文件上传接口",
                "parameters": [
                    {
                        "type": "file",
                        "description": "对象文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传编码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oss.BaseObject"
                        }
                    }
                }
            }
        },
        "/v1/oss/text/cover": {
            "post": {
                "description": "Oss覆盖文本上传接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oss文件上传接口"
                ],
                "summary": "Oss覆盖文本上传接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文本内容",
                        "name": "text",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传编码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "覆盖的源路径",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oss.BaseObject"
                        }
                    }
                }
            }
        },
        "/v1/oss/text/upload": {
            "post": {
                "description": "Oss文本上传接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oss文件上传接口"
                ],
                "summary": "Oss文本上传接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文本内容",
                        "name": "text",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传编码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oss.BaseObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpresult.Response": {
            "type": "object",
            "properties": {
                "bizCode": {
                    "description": "响应业务码",
                    "type": "string"
                },
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应信息",
                    "type": "string"
                },
                "success": {
                    "description": "响应是否成功",
                    "type": "boolean"
                },
                "ts": {
                    "description": "响应时间戳",
                    "type": "integer"
                }
            }
        },
        "oss.BaseObject": {
            "type": "object",
            "properties": {
                "bucket": {
                    "description": "对象存储bucket",
                    "type": "string"
                },
                "contentType": {
                    "description": "对象contentType",
                    "type": "string"
                },
                "fileName": {
                    "description": "对象源文件name",
                    "type": "string"
                },
                "format": {
                    "description": "对象格式，文件名后缀",
                    "type": "string"
                },
                "object": {
                    "description": "对象路径",
                    "type": "string"
                },
                "size": {
                    "description": "对象大小，单位Byte",
                    "type": "integer"
                },
                "uploadStatus": {
                    "description": "上传状态",
                    "type": "integer"
                },
                "url": {
                    "description": "真实对象地址",
                    "type": "string"
                }
            }
        }
    }
}